requirements.txt
discord.py==2.0.0
tortoise-orm==0.19.2
aerich==0.6.3
python-dotenv==0.20.0